// frontend/src/lib/types.ts
export interface RawLogEvent {
  timestamp?: number;
  message?: string;
  ingestionTime?: number;
  logStreamName?: string;
}
// ...
export interface LogGroupInfo {
  name: string;
  arn?: string;
  storedBytes?: number;
  creationTime?: number;
}